Core/requirements.txt
openai
rich
pydantic
python-dotenv
python-telegram-bot
